refac : new feature leave group

+++ src/component/Room.tsx
@@ -6,10 +6,10 @@ import firebase from "firebase/compat/app";
 import DocumentData = firebase.firestore.DocumentData;
 
 export function Room({id , room_name, createdBy} : RoomType | DocumentData){
-    const { deleteGroup, setRoomId, setChatRoom } = UseRoom()
+    const { deleteGroup, setRoomId, setChatRoom, leaveGroup } = UseRoom()
 
     function handleDelete(id : string){
-        deleteGroup(id)
+        confirm("Sure, want to delete the group?") &&  deleteGroup(id)
     }
 
     function handleRoomClick(id : string){
@@ -17,18 +17,23 @@ export function Room({id , room_name, createdBy} : RoomType | DocumentData){
         setChatRoom(room_name)
     }
 
+    function handleLeave(id:string){
+        confirm("Sure, want to leave the group?") && leaveGroup(id)
+    }
+
     return(
         <>
-            <Card  onClick={()=>handleRoomClick(id)}  className=" p-5 m-4 bg-gray-400 hover:cursor-pointer shadow-gray-400-xl bg-base-false " compact bordered>
+            <Card  onClick={()=>handleRoomClick(id)}  className=" p-3 m-4 bg-gray-400 hover:cursor-pointer shadow-gray-400-xl bg-base-false " compact bordered>
                <Card.Body className="text-neutral-400  flex flex-row justify-between items-center hover:cursor-pointer">
                        <div>
                             <Card.Title className=" font-bold" tag="h2">
                                 <h2>{room_name}</h2>
                             </Card.Title>
                             <p>Created By : {createdBy}</p>
+                           <Button className="btn mt-2 btn-neutral btn-sm overflow-hidden" onClick={()=>navigator.clipboard.writeText(`${id}`)}>copy Id</Button>
                         </div>
                        <div className="flex flex-col">
-                            <Button className="btn btn-ghost btn-sm" onClick={()=>navigator.clipboard.writeText(`${id}`)}>copy Id</Button>
+                           <button onClick={()=>handleLeave(id)} className="btn btn-ghost btn-sm overflow-hidden">Leave</button>
                            <button onClick={() => handleDelete(id)} className="btn btn-sm btn-ghost "><IoTrash style={{color: "red"}}/></button>
                        </div>
                 </Card.Body>
